comine the two forms
